+++ Basil/include/Renderer/Renderer.h
@@ -13,6 +13,8 @@ namespace Basil
 	{
 		public:
 			static void init();
+			static void onWindowResize(unsigned int width, unsigned int height);
+
 			static void beginScene(OrthographicCamera& camera);
 			static void endScene();
 			static void submit(const Shared<Shader>& shader, const Shared<VertexArray>& vao, const glm::mat4& transform = glm::mat4(1.0f));