+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change log for the Basil Game Engine project.
 ## 2021-04-15
 ### Added
 * An orthographic camera controller.
+* Proper window resize event handling.
 
 
 ## 2021-04-14